Moves Trait classes to separate directory.
