+++ example/GetNonPublicPropertyExample.php
@@ -14,7 +14,7 @@ class Example
 
 class ExampleTest extends \PHPUnit\Framework\TestCase
 {
-    use \Potherca\PhpUnit\GetNonPublicPropertyTrait;
+    use \Potherca\PhpUnit\Traits\GetNonPublicPropertyTrait;
 
     const MOCK_VALUE = 'mock-value';
