+++ example/PHP5.3/CreateDataProviderExample.php
@@ -7,7 +7,7 @@ class ExampleTest extends \PHPUnit\Framework\TestCase
     /**
      * As PHP5.3 does not support traits, __call is (a)bused instead of the trait.
      *
-     use \Potherca\PhpUnit\CreateDataProviderTrait;
+     use \Potherca\PhpUnit\Traits\CreateDataProviderTrait;
      *
      * @param string $name
      * @param array $parameters
