+++ example/PHP5.3/SetNonPublicPropertyExample.php
@@ -17,7 +17,7 @@ class ExampleTest extends \PHPUnit\Framework\TestCase
     /**
      * As PHP5.3 does not support traits, __call is (a)bused instead of the trait.
      *
-     use \Potherca\PhpUnit\SetNonPublicPropertyTrait;
+     use \Potherca\PhpUnit\Traits\SetNonPublicPropertyTrait;
      *
      * @param string $name
      * @param array $parameters
