+++ example/SetNonPublicPropertyExample.php
@@ -14,7 +14,7 @@ class Example
 
 class ExampleTest extends \PHPUnit\Framework\TestCase
 {
-    use \Potherca\PhpUnit\SetNonPublicPropertyTrait;
+    use \Potherca\PhpUnit\Traits\SetNonPublicPropertyTrait;
 
     const MOCK_VALUE = 'mock-value';
